browser/components/resistfingerprinting/test/mochitest/test_device_sensor_event.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1369319
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1369319</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1369319 **/
  SimpleTest.waitForExplicitFinish();
  window.onload = () => {
    SimpleTest.waitForFocus(() => {
      SpecialPowers.pushPrefEnv({"set":
        [
          ["device.sensors.test.events", true],
          ["privacy.resistFingerprinting", true]
        ]
      }, doTest);
    }, window);
  };

  function doTest() {
    window.addEventListener("devicemotion", () => {
      ok(false, "The device motion event should not be fired.");
    }, {once: true});

    window.addEventListener("TestEvent", () => {
      // If we receive this event without receiving a 'devicemotion' event, this means
      // the device sensor event has been blocked correctly.
      ok(true, "Got the 'TestEvent' event.");
      SimpleTest.finish();
    }, {once: true});

    window.dispatchEvent(new CustomEvent("TestEvent"));
  }
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test"></pre>
</body>
</html>
